revert: "feat: change catalog identifier from i32 to u64"
revert: This reverts commit 754a900cdf489f4bbbd74fbb430e23ad1381a598.

+++ rust/common/src/catalog/mod.rs
@@ -9,11 +9,11 @@ use pb_convert_derive::FromProtobuf;
 #[derive(Clone, Debug, Default, Hash, PartialOrd, PartialEq, Eq, FromProtobuf)]
 #[pb_convert(pb_type = "risingwave_proto::plan::DatabaseRefId")]
 pub struct DatabaseId {
-    database_id: u64,
+    database_id: i32,
 }
 
 impl DatabaseId {
-    pub fn new(database_id: u64) -> Self {
+    pub fn new(database_id: i32) -> Self {
         DatabaseId { database_id }
     }
 }
@@ -22,11 +22,11 @@ impl DatabaseId {
 #[pb_convert(pb_type = "risingwave_proto::plan::SchemaRefId")]
 pub struct SchemaId {
     database_ref_id: DatabaseId,
-    schema_id: u64,
+    schema_id: i32,
 }
 
 impl SchemaId {
-    pub fn new(database_ref_id: DatabaseId, schema_id: u64) -> Self {
+    pub fn new(database_ref_id: DatabaseId, schema_id: i32) -> Self {
         SchemaId {
             database_ref_id,
             schema_id,
@@ -38,11 +38,11 @@ impl SchemaId {
 #[pb_convert(pb_type = "risingwave_proto::plan::TableRefId")]
 pub struct TableId {
     schema_ref_id: SchemaId,
-    table_id: u64,
+    table_id: i32,
 }
 
 impl TableId {
-    pub fn new(schema_ref_id: SchemaId, table_id: u64) -> Self {
+    pub fn new(schema_ref_id: SchemaId, table_id: i32) -> Self {
         TableId {
             schema_ref_id,
             table_id,
